fix: Define SecretStoreAuth in separate file
Define SecretStoreAuthenticationHandlerFunc in separate file
