+++ bootstrap/container/clients.go
@@ -1,6 +1,6 @@
 //
 // Copyright (c) 2022 Intel Corporation
-// Copyright (C) 2024 IOTech Ltd
+// Copyright (C) 2024-2025 IOTech Ltd
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
