+++ bootstrap/handlers/auth_func.go
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright 2024 IOTech Ltd
+ * Copyright 2024-2025 IOTech Ltd
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
