+++ bootstrap/handlers/auth_middleware_no_ziti.go
@@ -1,7 +1,7 @@
 //go:build no_openziti
 
 /*******************************************************************************
- * Copyright 2024 IOTech Ltd
+ * Copyright 2024-2025 IOTech Ltd
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
@@ -23,9 +23,7 @@ import (
 
 	"github.com/edgexfoundry/go-mod-bootstrap/v4/bootstrap/container"
 	"github.com/edgexfoundry/go-mod-bootstrap/v4/bootstrap/handlers/headers"
-	"github.com/edgexfoundry/go-mod-bootstrap/v4/bootstrap/interfaces"
 	"github.com/edgexfoundry/go-mod-bootstrap/v4/di"
-	"github.com/edgexfoundry/go-mod-core-contracts/v4/clients/logger"
 	dtoCommon "github.com/edgexfoundry/go-mod-core-contracts/v4/dtos/common"
 
 	"github.com/golang-jwt/jwt/v5"
@@ -98,25 +96,4 @@ func AuthenticationHandlerFunc(dic *di.Container) echo.MiddlewareFunc {
 			return echo.NewHTTPError(http.StatusUnauthorized, http.StatusText(http.StatusUnauthorized))
 		}
 	}
-}
-
-// SecretStoreAuthenticationHandlerFunc verifies the JWT with a OpenBao-based JWT authentication check
-func SecretStoreAuthenticationHandlerFunc(secretProvider interfaces.SecretProviderExt, lc logger.LoggingClient, token string, c echo.Context) error {
-	r := c.Request()
-	w := c.Response()
-
-	validToken, err := secretProvider.IsJWTValid(token)
-	if err != nil {
-		lc.Errorf("Error checking JWT validity by the secret provider: %v ", err)
-		// set Response.Committed to true in order to rewrite the status code
-		w.Committed = false
-		return echo.NewHTTPError(http.StatusInternalServerError, http.StatusText(http.StatusInternalServerError))
-	} else if !validToken {
-		lc.Warnf("Request to '%s' UNAUTHORIZED", r.URL.Path)
-		// set Response.Committed to true in order to rewrite the status code
-		w.Committed = false
-		return echo.NewHTTPError(http.StatusUnauthorized, http.StatusText(http.StatusUnauthorized))
-	}
-	lc.Debugf("Request to '%s' authorized", r.URL.Path)
-	return nil
 }
