+++ bootstrap/handlers/constants.go
@@ -1,0 +1,9 @@
+//
+// Copyright (C) 2025 IOTech Ltd
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package handlers
+
+// openBaoIssuer defines the issuer if JWT was issued from OpenBao
+const openBaoIssuer = "/v1/identity/oidc"
